feat: ItemDTO 및 Item 객체 수정
ItemDTO의 역할에 맞게 불변성을 가지도록 수정
Item객체는 값 자체를 표현하도록 수정(vo)
