+++ src/main/java/gift/Model/Item.java
@@ -1,10 +1,30 @@
 package gift.Model;
 
+import java.util.Objects;
+
 public class Item {
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public void setPrice(Long price) {
+        this.price = price;
+    }
+
+    public void setImgUrl(String imgUrl) {
+        this.imgUrl = imgUrl;
+    }
+
     Long id;
     String name;
     Long price;
     String imgUrl;
+    public Item(){}
 
     public Item(Long id, String name, Long price, String imgUrl) {
         this.id = id;
@@ -24,5 +44,24 @@ public class Item {
 
     public String getImgUrl() {
         return imgUrl;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Item item = (Item) o;
+        return Objects.equals(id, item.id) && Objects.equals(name, item.name)
+            && Objects.equals(price, item.price) && Objects.equals(imgUrl,
+            item.imgUrl);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, name, price, imgUrl);
     }
 }
