+++ src/main/resources/application.properties
@@ -6,5 +6,5 @@ spring.thymeleaf.suffix=.html
 spring.mvc.hiddenmethod.filter.enabled=true
 spring.h2.console.enabled=true
 spring.datasource.username=sa
-spring.datasource.url=jdbc:h2:mem:test
+spring.datasource.url=jdbc:h2:mem:test2
 logging.level.org.springframework.jdbc=debug
